chore: restructure cli tests and add orgs-v5 smoke test
