+++ packages/cli/test/acceptance/plugin.acceptance.test.ts
